Complete hoomd.tuner -> hoomd.tune renaming
