+++ hoomd/__init__.py
@@ -74,7 +74,7 @@ from hoomd.state import State
 from hoomd.operations import Operations
 from hoomd.snapshot import Snapshot
 from hoomd.box import Box
-from hoomd import tuner
+from hoomd import tune
 from hoomd import output
 from hoomd import logging
 from hoomd import custom
