+++ hoomd/tune/__init__.py
@@ -1,3 +1,3 @@
-from hoomd.tuner.sorter import ParticleSorter
-from hoomd.tuner.balance import LoadBalancer
-from hoomd.tuner.custom_tuner import CustomTuner, _InternalCustomTuner
+from hoomd.tune.sorter import ParticleSorter
+from hoomd.tune.balance import LoadBalancer
+from hoomd.tune.custom_tuner import CustomTuner, _InternalCustomTuner
